Rename CollisionHandler and Computer to their proper name.
< organisation, naming
